Optimize setting element text for most common case
This avoids the performance cost of jQuery's text() method when called as a setter, using a new setText() method. The new method first checks whether the result set consists of a single element with only a single text node as child, in which case it sets the text node's nodeValue property directly. All other cases are forwarded on to the original method. The optimized case also makes an assumption that cleanData does not need to be called for the single element.